Merger now flattens models before merging if one model includes functional submodels and one doesn't. This removes the FunctionalSubmodel structures, and removes all MappableVariables that are inputs for their submodel. It also removes prefixes from the names of all data structures in the model.

+++ src/semsim/model/collection/SemSimModel.java
@@ -383,8 +383,12 @@ public class SemSimModel extends SemSimCollection implements Annotatable  {
 		
 	/**@return The parent FunctionalSubmodel for a MappableVariable.*/
 	public FunctionalSubmodel getParentFunctionalSubmodelForMappableVariable(MappableVariable var){
-		String compname = var.getName().substring(0, var.getName().lastIndexOf("."));
-		return (FunctionalSubmodel) getSubmodel(compname);
+		
+		if(var.getName().contains(".")){
+			String compname = var.getName().substring(0, var.getName().lastIndexOf("."));
+			return (FunctionalSubmodel) getSubmodel(compname);
+		}
+		else return null;
 	}
 	
 	/**
@@ -804,7 +808,7 @@ public class SemSimModel extends SemSimCollection implements Annotatable  {
 				}
 				
 				FunctionalSubmodel fs = getParentFunctionalSubmodelForMappableVariable(mapv);
-				fs.removeVariableEquationFromMathML(mapv);
+				if(fs != null) fs.removeVariableEquationFromMathML(mapv);
 			}
 			
 			for(Submodel sub : getSubmodels()){
